Backend/requirements.txt
# Core Django & REST Framework
Django
djangorestframework
djangorestframework-simplejwt[crypto]
django-cors-headers

# Database & Cache
psycopg2-binary
redis

# Async Tasks
celery[redis]
flower

# AI Resume Screening (Critical for our project)
sentence-transformers
torch --index-url https://download.pytorch.org/whl/cpu

# Document Processing
pypdf
python-docx

# API & Documentation
drf-spectacular
django-filter

# Utilities & Security
django-environ
gunicorn
whitenoise

# Dev Tools
django-debug-toolbar
django-extensions